requirements.txt
Flask==3.1.3
openai
requests
beautifulsoup4
PyMuPDF
python-dotenv